+++ tools/widgets/editor_viewport.vala
@@ -93,6 +93,9 @@ public class EditorViewport : Gtk.Bin
 		string args[] =
 		{
 			ENGINE_EXE,
+#if CROWN_PLATFORM_LINUX
+			"--display-server", "x11",
+#endif
 			"--data-dir",
 			_project.data_dir(),
 			"--boot-dir",